Fix updating mesh node when responding to unsolicited messages
In NodeConfigurationRepository the node would get updated when
the configuration state changed. This also happened when
the app received publish messages from multiple nodes. The result was that the
selected node for configuration was replaced under the hood and the user
configured a different node.

This patch fixes this by ignoring configuration state changes that do not belong
to the selected node.